refactor: Move email settings from General to Detailed
